refactored and cleaned up
